[timer] Avoid overflow of 32-bit time_t value by converting it to UInt64 before multiplication

+++ common/misc/timer.cc
@@ -60,7 +60,7 @@ UInt64 Timer::now()
 {
    timespec t;
    clock_gettime(CLOCK_REALTIME, &t);
-   return (t.tv_sec * 1000000000) + t.tv_nsec;
+   return (UInt64(t.tv_sec) * 1000000000) + t.tv_nsec;
 }
 
 void Timer::start()
